fixup 14169f876072567fc9d0341ba03bb9c62725986e

+++ src/util/DivisionModels.h
@@ -191,7 +191,7 @@ public:
 
 	const DivModelType& getRandomModel() const
 	{
-		const int id = (int)GetRandomUniform(0,models.size()-0.001);
+		const int id = (int)GetRandomUniform(0,(float)models.size()-0.001f);
 		return getModel(id);
 	}
